+++ e2e/fluid-grid.e2e-spec.ts
@@ -1,6 +1,7 @@
 import {
   expect,
-  SkyHostBrowser
+  SkyHostBrowser,
+  SkyVisualThemeSelector
 } from '@skyux-sdk/e2e';
 
 import {
@@ -9,112 +10,279 @@ import {
 } from 'protractor';
 
 describe('Fluid Grid', () => {
-  beforeEach(() => {
-    SkyHostBrowser.get('visual/fluid-grid');
+  let currentTheme: string;
+  let currentThemeMode: string;
+
+  async function scrollToTop(): Promise<void> {
+    await SkyHostBrowser.scrollTo('body');
+  }
+
+  async function selectTheme(theme: string, mode: string): Promise<void> {
+    currentTheme = theme;
+    currentThemeMode = mode;
+
+    await SkyVisualThemeSelector.selectTheme(theme, mode);
+    await scrollToTop();
+  }
+
+  function getScreenshotName(name: string): string {
+    if (currentTheme) {
+      name += '-' + currentTheme;
+    }
+
+    if (currentThemeMode) {
+      name += '-' + currentThemeMode;
+    }
+
+    return name;
+  }
+
+  beforeEach(async () => {
+    await SkyHostBrowser.get('visual/fluid-grid');
   });
 
-  it('should handle very small screens', (done) => {
-    SkyHostBrowser.setWindowDimensions(600, 800);
+  async function validateXs(done: DoneFn): Promise<void> {
+    await SkyHostBrowser.setWindowDimensions(600, 800);
+
     expect('#screenshot-fluid-grid-xsmall').toMatchBaselineScreenshot(done, {
-      screenshotName: 'fluid-grid-xs'
+      screenshotName: getScreenshotName('fluid-grid-xs')
     });
-  });
+  }
 
-  it('should display all columns evenly split on the same row when on a small screen', (done) => {
-    SkyHostBrowser.setWindowDimensions(800, 800);
+  async function validateEvenSplitSm(done: DoneFn): Promise<void> {
+    await SkyHostBrowser.setWindowDimensions(800, 800);
+
     expect('#screenshot-fluid-grid').toMatchBaselineScreenshot(done, {
-      screenshotName: 'fluid-grid-sm-evenrow'
+      screenshotName: getScreenshotName('fluid-grid-sm-evenrow')
     });
-  });
+  }
 
-  it('should display two on the top row and one on the bottom row when on a medium screen', (done) => {
-    SkyHostBrowser.setWindowDimensions(1100, 800);
+  async function validateTwoTopOneBottomMd(done: DoneFn): Promise<void> {
+    await SkyHostBrowser.setWindowDimensions(1100, 800);
+
     expect('#screenshot-fluid-grid').toMatchBaselineScreenshot(done, {
-      screenshotName: 'fluid-grid-md-tworows'
+      screenshotName: getScreenshotName('fluid-grid-md-tworows')
     });
-  });
+  }
 
-  it('should display all columns distributed unevenly when on a large screen', (done) => {
-    SkyHostBrowser.setWindowDimensions(1400, 800);
+  async function validateUnevenLg(done: DoneFn): Promise<void> {
+    await SkyHostBrowser.setWindowDimensions(1400, 800);
+
     expect('#screenshot-fluid-grid').toMatchBaselineScreenshot(done, {
-      screenshotName: 'fluid-grid-lg-unevenrow'
+      screenshotName: getScreenshotName('fluid-grid-lg-unevenrow')
     });
-  });
+  }
 
-  it('should reverse column order with reverseColumnOrder applied', (done) => {
-    SkyHostBrowser.setWindowBreakpoint('lg');
+  async function validateReverse(done: DoneFn): Promise<void> {
+    await SkyHostBrowser.setWindowBreakpoint('lg');
+
     expect('#screenshot-fluid-grid-reverse').toMatchBaselineScreenshot(done, {
-      screenshotName: 'fluid-grid-reversecolumnorder'
+      screenshotName: getScreenshotName('fluid-grid-reversecolumnorder')
     });
-  });
-});
+  }
 
-describe('Fluid Grid with wrapping "sky-fluid-grid" component', () => {
-  beforeEach(() => {
-    SkyHostBrowser.get('visual/fluid-grid');
-  });
+  async function validateWrapperXs(done: DoneFn): Promise<void> {
+    await SkyHostBrowser.setWindowDimensions(600, 800);
+
+    expect('#screenshot-fluid-grid-wrapper-xsmall').toMatchBaselineScreenshot(done, {
+      screenshotName: getScreenshotName('fluid-grid-wrapper-xs')
+    });
+  }
+
+  async function validateWrapperEvenSplitSm(done: DoneFn): Promise<void> {
+    await SkyHostBrowser.setWindowDimensions(800, 800);
+
+    expect('#screenshot-fluid-grid-wrapper').toMatchBaselineScreenshot(done, {
+      screenshotName: getScreenshotName('fluid-grid-wrapper-sm-evenrow')
+    });
+  }
+
+  async function validateWrapperTwoTopOneBottomMd(done: DoneFn): Promise<void> {
+    await SkyHostBrowser.setWindowDimensions(1100, 800);
+
+    expect('#screenshot-fluid-grid-wrapper').toMatchBaselineScreenshot(done, {
+      screenshotName: getScreenshotName('fluid-grid-wrapper-md-tworows')
+    });
+  }
+
+  async function validateWrapperUnevenLg(done: DoneFn): Promise<void> {
+    await SkyHostBrowser.setWindowDimensions(1400, 800);
+
+    expect('#screenshot-fluid-grid-wrapper').toMatchBaselineScreenshot(done, {
+      screenshotName: getScreenshotName('fluid-grid-wrapper-lg-unevenrow')
+    });
+  }
+
+  async function validateWrapperReverse(done: DoneFn): Promise<void> {
+    await SkyHostBrowser.setWindowBreakpoint('lg');
+
+    expect('#screenshot-fluid-grid-wrapper-reverse').toMatchBaselineScreenshot(done, {
+      screenshotName: getScreenshotName('fluid-grid-wrapper-reversecolumnorder')
+    });
+  }
+
+  async function validateRemoveMargins(done: DoneFn): Promise<void> {
+    await element.all(by.css('#disableMargin')).click();
+
+    expect('#screenshot-fluid-grid-gutters').toMatchBaselineScreenshot(done, {
+      screenshotName: getScreenshotName('screenshot-fluid-grid-remove-margins')
+    });
+  }
+
+  async function validateGuttersMedium(done: DoneFn): Promise<void> {
+    await element(by.cssContainingText('option', 'Medium')).click();
+
+    expect('#screenshot-fluid-grid-gutters').toMatchBaselineScreenshot(done, {
+      screenshotName: getScreenshotName('screenshot-fluid-grid-gutters-medium')
+    });
+  }
+
+  async function validateGuttersSmall(done: DoneFn): Promise<void> {
+    await element(by.cssContainingText('option', 'Small')).click();
+
+    expect('#screenshot-fluid-grid-gutters').toMatchBaselineScreenshot(done, {
+      screenshotName: getScreenshotName('screenshot-fluid-grid-gutters-small')
+    });
+  }
 
   it('should handle very small screens', (done) => {
-    SkyHostBrowser.setWindowDimensions(600, 800);
-    expect('#screenshot-fluid-grid-wrapper-xsmall').toMatchBaselineScreenshot(done, {
-      screenshotName: 'fluid-grid-wrapper-xs'
-    });
+    validateXs(done);
   });
 
   it('should display all columns evenly split on the same row when on a small screen', (done) => {
-    SkyHostBrowser.setWindowDimensions(800, 800);
-    expect('#screenshot-fluid-grid-wrapper').toMatchBaselineScreenshot(done, {
-      screenshotName: 'fluid-grid-wrapper-sm-evenrow'
-    });
+    validateEvenSplitSm(done);
   });
 
-  it('should display two on the top row and one on the bottom row when on a medium screen', (done) => {
-    SkyHostBrowser.setWindowDimensions(1100, 800);
-    expect('#screenshot-fluid-grid-wrapper').toMatchBaselineScreenshot(done, {
-      screenshotName: 'fluid-grid-wrapper-md-tworows'
-    });
-  });
+  it(
+    'should display two on the top row and one on the bottom row when on a medium screen',
+    (done) => {
+      validateTwoTopOneBottomMd(done);
+    }
+  );
 
   it('should display all columns distributed unevenly when on a large screen', (done) => {
-    SkyHostBrowser.setWindowDimensions(1400, 800);
-    expect('#screenshot-fluid-grid-wrapper').toMatchBaselineScreenshot(done, {
-      screenshotName: 'fluid-grid-wrapper-lg-unevenrow'
-    });
+    validateUnevenLg(done);
   });
 
   it('should reverse column order with reverseColumnOrder applied', (done) => {
-    SkyHostBrowser.setWindowBreakpoint('lg');
-    expect('#screenshot-fluid-grid-wrapper-reverse').toMatchBaselineScreenshot(done, {
-      screenshotName: 'fluid-grid-wrapper-reversecolumnorder'
+    validateReverse(done);
+  });
+
+  describe('with wrapping "sky-fluid-grid" component', () => {
+    it('should handle very small screens', (done) => {
+      validateWrapperXs(done);
+    });
+
+    it('should display all columns evenly split on the same row when on a small screen', (done) => {
+      validateWrapperEvenSplitSm(done);
+    });
+
+    it(
+      'should display two on the top row and one on the bottom row when on a medium screen',
+      (done) => {
+        validateWrapperTwoTopOneBottomMd(done);
+      }
+    );
+
+    it('should display all columns distributed unevenly when on a large screen', (done) => {
+      validateWrapperUnevenLg(done);
+    });
+
+    it('should reverse column order with reverseColumnOrder applied', (done) => {
+      validateWrapperReverse(done);
+    });
+
+    describe('with adjustable gutters/margins', () => {
+      beforeEach(async () => {
+        await SkyHostBrowser.setWindowBreakpoint('lg');
+        await SkyHostBrowser.scrollTo('#screenshot-fluid-grid-gutters');
+      });
+
+      it('should remove margins', (done) => {
+        validateRemoveMargins(done);
+      });
+
+      it('should change gutters to medium', (done) => {
+        validateGuttersMedium(done);
+      });
+
+      it('should change gutters to small', (done) => {
+        validateGuttersSmall(done);
+      });
     });
   });
 
-  describe('Fluid Grid with adjustable gutters/margins', () => {
-    beforeEach(() => {
-      SkyHostBrowser.get('visual/fluid-grid');
-      SkyHostBrowser.setWindowBreakpoint('lg');
-      SkyHostBrowser.scrollTo('#screenshot-fluid-grid-gutters');
+  describe('when modern theme', () => {
+    beforeEach(async () => {
+      await selectTheme('modern', 'light');
     });
 
-    it('should remove margins', (done) => {
-      element.all(by.css('#disableMargin')).click();
-      expect('#screenshot-fluid-grid-gutters').toMatchBaselineScreenshot(done, {
-        screenshotName: 'screenshot-fluid-grid-remove-margins'
+    it('should handle very small screens', (done) => {
+      validateXs(done);
+    });
+
+    it('should display all columns evenly split on the same row when on a small screen', (done) => {
+      validateEvenSplitSm(done);
+    });
+
+    it(
+      'should display two on the top row and one on the bottom row when on a medium screen',
+      (done) => {
+        validateTwoTopOneBottomMd(done);
+      }
+    );
+
+    it('should display all columns distributed unevenly when on a large screen', (done) => {
+      validateUnevenLg(done);
+    });
+
+    it('should reverse column order with reverseColumnOrder applied', (done) => {
+      validateReverse(done);
+    });
+
+    describe('with wrapping "sky-fluid-grid" component', () => {
+      it('should handle very small screens', (done) => {
+        validateWrapperXs(done);
       });
-    });
 
-    it('should change gutters to medium', (done) => {
-      element(by.cssContainingText('option', 'Medium')).click();
-      expect('#screenshot-fluid-grid-gutters').toMatchBaselineScreenshot(done, {
-        screenshotName: 'screenshot-fluid-grid-gutters-medium'
+      it('should display all columns evenly split on the same row when on a small screen', (done) => {
+        validateWrapperEvenSplitSm(done);
       });
-    });
 
-    it('should change gutters to small', (done) => {
-      element(by.cssContainingText('option', 'Small')).click();
-      expect('#screenshot-fluid-grid-gutters').toMatchBaselineScreenshot(done, {
-        screenshotName: 'screenshot-fluid-grid-gutters-small'
+      it(
+        'should display two on the top row and one on the bottom row when on a medium screen',
+        (done) => {
+          validateWrapperTwoTopOneBottomMd(done);
+        }
+      );
+
+      it('should display all columns distributed unevenly when on a large screen', (done) => {
+        validateWrapperUnevenLg(done);
+      });
+
+      it('should reverse column order with reverseColumnOrder applied', (done) => {
+        validateWrapperReverse(done);
+      });
+
+      describe('with adjustable gutters/margins', () => {
+        beforeEach(async () => {
+          await SkyHostBrowser.setWindowBreakpoint('lg');
+          await SkyHostBrowser.scrollTo('#screenshot-fluid-grid-gutters');
+        });
+
+        it('should remove margins', (done) => {
+          validateRemoveMargins(done);
+        });
+
+        it('should change gutters to medium', (done) => {
+          validateGuttersMedium(done);
+        });
+
+        it('should change gutters to small', (done) => {
+          validateGuttersSmall(done);
+        });
       });
     });
   });
+
 });