+++ src/app/public/modules/action-button/action-button.component.ts
@@ -7,17 +7,26 @@ import {
 
 import {
   SkyActionButtonPermalink
-} from './types';
+} from './action-button-permalink';
 
+/**
+ * Creates a button to present users with an option to move forward with tasks.
+ */
 @Component({
   selector: 'sky-action-button',
   styleUrls: ['./action-button.component.scss'],
   templateUrl: './action-button.component.html'
 })
 export class SkyActionButtonComponent {
+/**
+ * Specifies a link for the action button.
+ */
   @Input()
   public permalink: SkyActionButtonPermalink;
 
+/**
+ * Fires when users select the action button.
+ */
   @Output()
   public actionClick = new EventEmitter<any>();