+++ src/app/public/modules/definition-list/definition-list.service.ts
@@ -1,7 +1,12 @@
 import { Injectable } from '@angular/core';
 
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
+import {
+  BehaviorSubject
+} from 'rxjs';
 
+/**
+ * @internal
+ */
 @Injectable()
 export class SkyDefinitionListService {
   public labelWidth = new BehaviorSubject<string>('');