+++ src/app/public/modules/page-summary/page-summary-key-info/page-summary-key-info.component.ts
@@ -1,8 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'sky-page-summary-key-info',
-  templateUrl: './page-summary-key-info.component.html',
-  styleUrls: ['./page-summary-key-info.component.scss']
-})
-export class SkyPageSummaryKeyInfoComponent { }