+++ src/app/public/modules/text-expand/fixtures/text-expand.component.fixture.ts
@@ -13,7 +13,10 @@ import {
 })
 export class TextExpandTestComponent {
 
-  @ViewChild(SkyTextExpandComponent, { read: SkyTextExpandComponent })
+  @ViewChild(SkyTextExpandComponent, {
+    read: SkyTextExpandComponent,
+    static: false
+  })
   public textExpand: SkyTextExpandComponent;
   public text: string;
   public maxLength: number;